Move secret data type to its own file
